+++ core-methods/src/main/java/com/mercadopago/sdk/android/coremethods/data/datasource/remote/ThreeDSDataSource.kt
@@ -1,5 +1,6 @@
 package com.mercadopago.sdk.android.coremethods.data.datasource.remote
 
+import com.mercadopago.sdk.android.coremethods.data.remote.request.UpdateThreeDSChallengeStatusRequest
 import com.mercadopago.sdk.android.coremethods.domain.model.ResultError
 import com.mercadopago.sdk.android.coremethods.domain.model.ThreeDSChallengeAuthentication
 import com.mercadopago.sdk.android.coremethods.domain.utils.Result
@@ -8,4 +9,9 @@ internal interface ThreeDSDataSource {
     suspend fun authenticateThreeDSChallenge(
         challengeId: String,
     ): Result<ThreeDSChallengeAuthentication, ResultError>
+
+    suspend fun updateThreeDSChallengeStatus(
+        challengeId: String,
+        request: UpdateThreeDSChallengeStatusRequest,
+    ): Result<Unit, ResultError>
 }
