+++ core-methods/src/main/java/com/mercadopago/sdk/android/coremethods/data/remote/service/CoreMethodsService.kt
@@ -1,7 +1,6 @@
 package com.mercadopago.sdk.android.coremethods.data.remote.service
 
 import com.mercadopago.sdk.android.coremethods.data.remote.request.CardTokenBodyRequest
-import com.mercadopago.sdk.android.coremethods.data.remote.request.UpdateThreeDSChallengeStatusRequest
 import com.mercadopago.sdk.android.coremethods.data.remote.response.CardIssuerResponse
 import com.mercadopago.sdk.android.coremethods.data.remote.response.CardTokenResponse
 import com.mercadopago.sdk.android.coremethods.data.remote.response.IdentificationTypesResponse
@@ -12,9 +11,7 @@ import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.Header
-import retrofit2.http.PATCH
 import retrofit2.http.POST
-import retrofit2.http.Path
 import retrofit2.http.Query
 import java.math.BigDecimal
 
@@ -51,10 +48,4 @@ internal interface CoreMethodsService {
         @Query("product_id") productId: String? = PRODUCT_ID,
         @Query("bin") bin: Int?,
     ): Response<List<PaymentMethodResponse>>
-
-    @PATCH("$BRICKS_API/$VERSION/challenges/threeds/{challengeId}")
-    suspend fun updateThreeDSChallengeStatus(
-        @Path("challengeId") challengeId: String,
-        @Body request: UpdateThreeDSChallengeStatusRequest,
-    ): Response<Unit>
 }
