+++ core-methods/src/main/java/com/mercadopago/sdk/android/coremethods/data/remote/service/ThreeDSService.kt
@@ -1,7 +1,10 @@
 package com.mercadopago.sdk.android.coremethods.data.remote.service
 
+import com.mercadopago.sdk.android.coremethods.data.remote.request.UpdateThreeDSChallengeStatusRequest
 import com.mercadopago.sdk.android.coremethods.data.remote.response.ThreeDSChallengeAuthenticationResponse
 import retrofit2.Response
+import retrofit2.http.Body
+import retrofit2.http.PATCH
 import retrofit2.http.POST
 import retrofit2.http.Path
 
@@ -13,4 +16,10 @@ internal interface ThreeDSService {
     suspend fun authenticateThreeDSChallenge(
         @Path("challengeId") challengeId: String,
     ): Response<ThreeDSChallengeAuthenticationResponse>
+
+    @PATCH("$BRICKS_API/$VERSION/challenges/threeds/{challengeId}")
+    suspend fun updateThreeDSChallengeStatus(
+        @Path("challengeId") challengeId: String,
+        @Body request: UpdateThreeDSChallengeStatusRequest,
+    ): Response<Unit>
 }
