+++ core-methods/src/main/java/com/mercadopago/sdk/android/coremethods/data/repository/ThreeDSRepositoryImpl.kt
@@ -1,9 +1,11 @@
 package com.mercadopago.sdk.android.coremethods.data.repository
 
 import com.mercadopago.sdk.android.coremethods.data.datasource.remote.ThreeDSDataSource
+import com.mercadopago.sdk.android.coremethods.data.remote.mappers.toRequest
 import com.mercadopago.sdk.android.coremethods.domain.model.ResultError
 import com.mercadopago.sdk.android.coremethods.domain.model.ThreeDSChallengeAuthentication
 import com.mercadopago.sdk.android.coremethods.domain.model.params.AuthenticateThreeDSChallengeParams
+import com.mercadopago.sdk.android.coremethods.domain.model.params.UpdateThreeDSChallengeStatusParams
 import com.mercadopago.sdk.android.coremethods.domain.repository.ThreeDSRepository
 import com.mercadopago.sdk.android.coremethods.domain.utils.Result
 
@@ -14,5 +16,14 @@ internal class ThreeDSRepositoryImpl(
         params: AuthenticateThreeDSChallengeParams,
     ): Result<ThreeDSChallengeAuthentication, ResultError> {
         return dataSource.authenticateThreeDSChallenge(params.challengeId)
+    }
+
+    override suspend fun updateThreeDSChallengeStatus(
+        params: UpdateThreeDSChallengeStatusParams,
+    ): Result<Unit, ResultError> {
+        return dataSource.updateThreeDSChallengeStatus(
+            challengeId = params.challengeId,
+            request = params.toRequest(),
+        )
     }
 }
