+++ core-methods/src/main/java/com/mercadopago/sdk/android/coremethods/domain/repository/CoreMethodsRepository.kt
@@ -10,7 +10,6 @@ import com.mercadopago.sdk.android.coremethods.domain.model.params.GenerateCardT
 import com.mercadopago.sdk.android.coremethods.domain.model.params.GetCardIssuersParams
 import com.mercadopago.sdk.android.coremethods.domain.model.params.GetInstallmentParams
 import com.mercadopago.sdk.android.coremethods.domain.model.params.GetPaymentMethodsParams
-import com.mercadopago.sdk.android.coremethods.domain.model.params.UpdateThreeDSChallengeStatusParams
 import com.mercadopago.sdk.android.coremethods.domain.utils.Result
 
 internal interface CoreMethodsRepository {
@@ -31,8 +30,4 @@ internal interface CoreMethodsRepository {
     suspend fun getPaymentMethods(
         params: GetPaymentMethodsParams,
     ): Result<List<PaymentMethod>, ResultError>
-
-    suspend fun updateThreeDSChallengeStatus(
-        params: UpdateThreeDSChallengeStatusParams,
-    ): Result<Unit, ResultError>
 }
